BREAKING CHANGE: Refactored FSID
Sometimes the FSID was used as a string, sometimes as a []byte, which is
generally fine except that as a []byte it sometimes included the UUID
dashes and sometimes did not. Refactored so all uses of FSID use
strings, except at that last moment where a []byte is required (for
Writes, etc.) in which case the []byte(stringvalue) is always used,
which includes the dashes. This gives a side affect that we can use
non-UUID file system identifiers should we choose to do so.